requirements.txt
dj-database-url==0.4.2
Django==1.10.5
django-allauth==0.25.2
django-annoying==0.10.3
django-bower==5.0.4
django-datetime-widget==0.9.3
django-scheduler==0.8
gunicorn==19.6.0
icalendar==3.11.1
oauthlib==2.0.1
psycopg2==2.6.2
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2016.10
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
whitenoise==3.2.3
